[LayoutNG] OOFs and column balancing
OOFs currently only affect the initial column balancing pass, but we
want them to affect column balancing on later passes, too. To do
so, we need to lay out the OOFs during column layout. Currently,
though, OOFs are laid out at the end of multicol layout (to ensure
the containing block has completed layout).

However, we can perform layout earlier for OOFs whose CB hasn't broken
so that they affect column balancing. The problem is that these
columns haven't been added to the builder at the time when the
columns are being balanced, so the OOFs haven't been propagated up
to the builder, either.

To address this, add a new ColumnBalancingInfo struct to hold the
current columns so that NGOutOfFlowLayoutPart will know which
columns to use. This will also store a list of OOFs so that we
can propagate the OOFs to this struct rather than to the builder.

This struct will then be passed into HandleFragmentation() so that
a special layout pass can be run on the current unbalanced columns,
where we will propagate the info needed to successfully
affect column balancing (such as the minimal space shortage,
# of new columns added, and if there was a violating break). This
special layout pass won't update the builder (i.e. the OOF results
aren't added and new columns aren't created).

Since we don't actually add the OOFs in this pass to the builder,
we will continue to lay out the OOFs at a later point (as is
currently done) where they will be officially laid out and added
to the builder.

Note that out-of-flow-in-multicolumn-099.html passed previously to
this CL since break-inside:avoid doesn't affect the OOF. This was
simply added to make sure that case was working as expected.

Bug: 1341358
Change-Id: Ic1fa6b57a8e4fa26a3f0e1acbd2efe7425787ab0