Add CSS properties to control pop-up hover behaviors
This CL adds these two CSS properties (which are still subject to
bikeshedding on the name):
 - hover-pop-up-delay
 - hover-pop-up-hide-delay

Both properties are time valued properties.  The first controls the
delay after hovering an invoking element (using the hoverpopup
attribute) before the pop-up is triggered. The second controls the
delay after which a showing pop-up will be hidden if it isn't hovered
for a period of time. The first is connected to the behavior, the
second will be connected in a followup CL.

Bug: 1307772
Change-Id: Ia44cbf63f91732935d620c527265129ffeb4f96b