Remove duplication in ConsumeCustomIdentWithToken.
There was a small amount of code duplication (both in code size and
runtime execution) between ConsumeCustomIdent and
ConsumeCustomIdentWithToken.  However, the latter had one caller that
can easily be adjusted to use the former.

This also adds a test to exercise parsing of paint() functions since I
couldn't find such a test.

Change-Id: I0c55ac1c174a1bb8ede85b6b1fbfc3369a80b4fc