Follow containing block chain when finding scroll container for sticky
When a sticky position element is contained by a fixed or absolute
position element, the scroll container of the sticky position element
should follow the containing block chain of the fixed or absolute
position element. Previously we followed the DOM hierarchy, causing
weird behavior in the case, e.g. the sticky element moves in the
opposite direction when the non-containing ancestor scroller scrolls.

Bug: 881555
Change-Id: I11994dfea2910283f06662558a8445362f15bca7